move fake db code into its own file
