+++ postgresql-cdc/src/test/scala/akka/stream/alpakka/postgresqlcdc/PostgreSQLCapturerSpec.scala
@@ -4,15 +4,14 @@
 
 package akka.stream.alpakka.postgresqlcdc
 
-import java.sql.{Connection, DriverManager}
+import java.sql.Connection
 
 import akka.actor.ActorSystem
 import akka.event.Logging
-import akka.stream.{ActorMaterializer, Attributes}
 import akka.stream.alpakka.postgresqlcdc.scaladsl._
 import akka.stream.testkit.scaladsl.TestSink
+import akka.stream.{ActorMaterializer, Attributes}
 import akka.testkit.{ImplicitSender, TestKit}
-import org.postgresql.util.PGobject
 import org.scalatest.{BeforeAndAfterAll, Matchers, WordSpecLike}
 
 abstract class PostgreSQLCapturerSpec(postgreSQLPortNumber: Int)
@@ -21,6 +20,8 @@ abstract class PostgreSQLCapturerSpec(postgreSQLPortNumber: Int)
     with ImplicitSender
     with Matchers
     with BeforeAndAfterAll {
+
+  import FakeDb._
 
   private val log = Logging(system, classOf[PostgreSQLCapturerSpec])
 
@@ -31,169 +32,9 @@ abstract class PostgreSQLCapturerSpec(postgreSQLPortNumber: Int)
 
   private implicit val conn: Connection = getConnection(connectionString)
 
-  private def getConnection(connectionString: String): Connection = {
-    val driver = "org.postgresql.Driver"
-    Class.forName(driver)
-    DriverManager.getConnection(connectionString)
-  }
-
-  private def createCustomersTable()(implicit conn: Connection): Unit = {
-    val createStatement =
-      conn.prepareStatement("""
-          |CREATE TABLE customers (
-          |  id SERIAL,
-          |  first_name VARCHAR(255) NOT NULL,
-          |  last_name VARCHAR(255) NOT NULL,
-          |  email VARCHAR(255) NOT NULL,
-          |  PRIMARY KEY(id)
-          |);
-        """.stripMargin)
-
-    createStatement.execute()
-  }
-
-  private def createSalesTable()(implicit conn: Connection): Unit = {
-    val createStatement =
-      conn.prepareStatement("""
-          |CREATE TABLE sales (
-          | id SERIAL NOT NULL PRIMARY KEY,
-          | info JSONB NOT NULL
-          |);
-        """.stripMargin)
-    createStatement.execute()
-
-  }
-
-  private def createPurchaseOrdersTable()(implicit conn: Connection): Unit =
-    conn.prepareStatement("""
-      |CREATE TABLE purchase_orders (
-      | id SERIAL NOT NULL PRIMARY KEY,
-      | info XML NOT NULL
-      | );
-    """.stripMargin).execute()
-
-  private def createEmployeesTable()(implicit conn: Connection): Unit =
-    conn.prepareStatement("""
-        |CREATE TABLE employees (
-        | id serial NOT NULL PRIMARY KEY,
-        | name VARCHAR(255) NOT NULL,
-        | position VARCHAR(255) DEFAULT NULL
-        |);
-        |
-      """.stripMargin).execute()
-
-  private def dropTableCustomers()(implicit conn: Connection): Unit =
-    conn.prepareStatement("DROP TABLE customers;").execute()
-
-  private def dropTableSales()(implicit conn: Connection): Unit =
-    conn.prepareStatement("DROP TABLE sales;").execute()
-
-  private def dropTablePurchaseOrders()(implicit conn: Connection): Unit =
-    conn.prepareStatement("DROP TABLE purchase_orders;").execute()
-
-  private def dropTableEmployees()(implicit conn: Connection): Unit =
-    conn.prepareStatement("DROP TABLE employees;").execute()
-
-  private def insertCustomer(id: Int, fName: String, lName: String, email: String)(implicit conn: Connection): Unit = {
-    val insertStatement =
-      conn.prepareStatement("INSERT INTO customers(id, first_name, last_name, email) VALUES(?, ?, ?, ?)")
-    insertStatement.setInt(1, id)
-    insertStatement.setString(2, fName)
-    insertStatement.setString(3, lName)
-    insertStatement.setString(4, email)
-    insertStatement.execute()
-  }
-
-  private def updateCustomerEmail(id: Int, newEmail: String)(implicit conn: Connection): Unit = {
-    val updateStatement =
-      conn.prepareStatement("UPDATE customers SET email = ? WHERE Id = ?")
-    updateStatement.setString(1, newEmail)
-    updateStatement.setInt(2, id)
-    updateStatement.execute()
-  }
-
-  private def deleteCustomers()(implicit conn: Connection): Unit = {
-    val deleteStatement =
-      conn.prepareStatement("DELETE FROM customers")
-    deleteStatement.execute()
-  }
-
-  private def insertSale(id: Int, info: String)(implicit conn: Connection): Unit = {
-    val pgObject = new PGobject
-    pgObject.setType("jsonb")
-    pgObject.setValue(info)
-    val insertStatement =
-      conn.prepareStatement("INSERT INTO sales(id, info) VALUES (?, ?);")
-    insertStatement.setInt(1, id)
-    insertStatement.setObject(2, pgObject)
-    insertStatement.execute()
-  }
-
-  private def updateSale(id: Int, newInfo: String)(implicit conn: Connection): Unit = {
-    val pgObject = new PGobject
-    pgObject.setType("jsonb")
-    pgObject.setValue(newInfo)
-    val updateStatement =
-      conn.prepareStatement("UPDATE sales SET info = ? WHERE id = ?;")
-    updateStatement.setObject(1, pgObject)
-    updateStatement.setInt(2, id)
-    updateStatement.execute()
-  }
-
-  private def deleteSale(id: Int)(implicit conn: Connection): Unit = {
-    val deleteStatement =
-      conn.prepareStatement("DELETE FROM sales WHERE id = ?;")
-    deleteStatement.setInt(1, id)
-    deleteStatement.execute()
-  }
-
-  private def insertPurchaseOrder(id: Int, info: String)(implicit conn: Connection): Unit = {
-    val pGobject = new PGobject
-    pGobject.setType("XML")
-    pGobject.setValue(info)
-    val insertStatement = conn.prepareStatement("INSERT INTO purchase_orders(id, info) VALUES (?, ?);")
-    insertStatement.setInt(1, id)
-    insertStatement.setObject(2, pGobject)
-    insertStatement.execute()
-  }
-
-  private def deletePurchaseOrder(id: Int)(implicit conn: Connection): Unit = {
-    val deleteStatement =
-      conn.prepareStatement("DELETE FROM purchase_orders WHERE id = ?;")
-    deleteStatement.setInt(1, id)
-    deleteStatement.execute()
-  }
-
-  private def insertEmployee(id: Int, name: String, position: String)(implicit conn: Connection): Unit = {
-    val insertStatement =
-      conn.prepareStatement("INSERT INTO employees(id, name, position) VALUES(?, ?, ?);")
-    insertStatement.setInt(1, id)
-    insertStatement.setString(2, name)
-    insertStatement.setString(3, position)
-    insertStatement.execute()
-  }
-
-  private def updateEmployee(id: Int, newPosition: String)(implicit conn: Connection): Unit = {
-    val updateStatement =
-      conn.prepareStatement("UPDATE employees SET position = ? WHERE id = ?;")
-    updateStatement.setString(1, newPosition)
-    updateStatement.setInt(2, id)
-    updateStatement.execute()
-  }
-
-  private def setUpLogicalDecodingSlot()(implicit conn: Connection): Unit = {
-    val stmt = conn.prepareStatement("SELECT * FROM pg_create_logical_replication_slot('scalatest','test_decoding')")
-    stmt.execute()
-  }
-
-  private def dropLogicalDecodingSlot()(implicit conn: Connection): Unit = {
-    val stmt = conn.prepareStatement("SELECT * FROM pg_drop_replication_slot('scalatest')")
-    stmt.execute()
-  }
-
   override def beforeAll(): Unit = {
     log.info("setting up logical decoding slot and creating customers table")
-    setUpLogicalDecodingSlot()
+    setUpLogicalDecodingSlot("scalatest")
     createCustomersTable()
     createSalesTable()
     createPurchaseOrdersTable()
@@ -206,7 +47,7 @@ abstract class PostgreSQLCapturerSpec(postgreSQLPortNumber: Int)
 
     // The following are useful for local testing but not necessary when running this test on proper CI (like Travis) since
     // the CI creates fresh docker containers and destroys them after the test is complete anyway.
-    dropLogicalDecodingSlot()
+    dropLogicalDecodingSlot("scalatest")
     dropTableCustomers()
     dropTableSales()
     dropTablePurchaseOrders()
